Also made codon_mutations parallel
